Refactor connection logic into separate module
